requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
supabase==2.7.4
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0